packages/sources/lition/test/adapter.test.ts
import { assert } from 'chai'
import { Requester } from '@chainlink/ea-bootstrap'
import { assertSuccess, assertError } from '@chainlink/ea-test-helpers'
import { AdapterRequest } from '@chainlink/types'
import { makeExecute } from '../src/adapter'

describe('execute', () => {
  const jobID = '1'
  const execute = makeExecute()

  context('successful calls @integration', () => {
    const requests = [
      {
        name: 'id not supplied',
        testData: { data: { source: 1 } },
      },
      {
        name: 'is is supplied',
        testData: { id: jobID, data: { source: 1 } },
      },
      {
        name: 'with date',
        testData: {
          id: jobID,
          data: { source: 1, date: `${new Date().toISOString().slice(0, 10)}` },
        },
      },
      {
        name: 'with hour',
        testData: { id: jobID, data: { source: 1, hour: 0 } },
      },
    ]

    requests.forEach((req) => {
      it(`${req.name}`, async () => {
        const data = await execute(req.testData as AdapterRequest)
        assertSuccess({ expected: 200, actual: data.statusCode }, data, jobID)
        assert.isAbove(data.result, 0)
        assert.isAbove(data.data.result, 0)
      })
    })
  })

  context('validation error', () => {
    const requests = [
      { name: 'empty body', testData: {} },
      { name: 'source not supplied', testData: { data: {} } },
    ]

    requests.forEach((req) => {
      it(`${req.name}`, async () => {
        try {
          await execute(req.testData as AdapterRequest)
        } catch (error) {
          const errorResp = Requester.errored(jobID, error)
          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
        }
      })
    })
  })

  context('error calls @integration', () => {
    const requests = [
      {
        name: 'unknown source',
        testData: { id: jobID, data: { source: 'not_real' } },
      },
      {
        name: 'unknown date',
        testData: { id: jobID, data: { source: 1, date: 'not_real' } },
      },
      {
        name: 'unknown hour',
        testData: { id: jobID, data: { source: 1, hour: 'not_real' } },
      },
    ]

    requests.forEach((req) => {
      it(`${req.name}`, async () => {
        try {
          await execute(req.testData as AdapterRequest)
        } catch (error) {
          const errorResp = Requester.errored(jobID, error)
          assertError({ expected: 500, actual: errorResp.statusCode }, errorResp, jobID)
        }
      })
    })
  })
})
